+++ drivers/CMakeLists.txt
@@ -6,7 +6,8 @@ add_executable(batch_chem batch_chem.cc cfg_reader.cc demo_containers.cc
                           demo_output.cc demo_utils.cc string_tokenizer.cc)
 
 # Reactive transport (0D) driver.
-add_executable(transport transport.c ini.c)
+add_executable(transport TransportInput.c TransportDriver.c DriverOutput.c 
+               transport.c ini.c)
 
 # Install executable.
 install(TARGETS batch_chem transport DESTINATION bin)
